Added error checking when loading a lua script
