+++ src/map/include/systems/inventorysystem.h
@@ -3,6 +3,7 @@
 #include "system.h"
 
 #include "item.h"
+#include <optional>
 
 namespace RoseCommon {
 class CliEquipItem;
@@ -23,7 +24,7 @@ class InventorySystem : public System {
         static bool addItem(Entity entity, RoseCommon::Item&& item);
         static RoseCommon::Item removeItem(Entity entity, uint8_t slot);
 
-        RoseCommon::Item buildItem(uint8_t type, uint16_t id, uint16_t life = 1000, bool isAppraised = true);
+        std::optional<RoseCommon::Item> buildItem(uint8_t type, uint16_t id, uint16_t life = 1000, bool isAppraised = true);
 
         void processEquip(CMapClient& client, Entity entity, const RoseCommon::CliEquipItem &packet);
 };
