+++ src/map/src/systems/inventorysystem.cpp
@@ -106,16 +106,19 @@ Item InventorySystem::removeItem(Entity entity, uint8_t slot) {
     return item;
 }
 
-Item InventorySystem::buildItem(uint8_t type, uint16_t id, uint16_t life, bool isAppraised) {
+std::optional<Item> InventorySystem::buildItem(uint8_t type, uint16_t id, uint16_t life, bool isAppraised) {
     const auto &itemDb = ItemDatabase::getInstance();
     const auto def = itemDb.getItemDef(type, id);
     Item item{def};
     item.life_ = life;
     item.isAppraised_ = isAppraised;
     auto luaSystem = manager_.get<LuaSystem>();
-    item.lua_ = luaSystem->loadScript<RoseCommon::ItemAPI>(def.script);
+    auto lua = luaSystem->loadScript<RoseCommon::ItemAPI>(def.script);
+    if (!lua)
+        return {};
+    item.lua_ = std::move(lua.value());
     //const auto& env = item.lua_.getEnv();
     // TODO: get durability from lua env
     item.lua_.onInit();
-    return item;
+    return std::make_optional(std::move(item));
 }
