+++ src/rosecommon/include/entityapi.h
@@ -13,8 +13,6 @@ class EntityAPI : public LuaAPI {
          logger_->warn("lua display call: {}", data);
       });
     }
-
-    void onInit() { auto f = env_["onInit"]; f(); }
 };
 
 }
