+++ src/rosecommon/include/luaapi.h
@@ -4,6 +4,7 @@
 #include "logconsole.h"
 #include <memory>
 #include "throwassert.h"
+#include <optional>
 
 namespace RoseCommon {
 
@@ -23,10 +24,26 @@ class LuaAPI {
             return env_;
         }
 
-        bool isCreated_;
     protected:
+        bool isCreated_;
         sol::environment env_;
         std::shared_ptr<spdlog::logger> logger_;
+
+        template <typename T, typename... Args>
+        std::optional<T> safeLuaCall(const std::string& func, const Args&... args) {
+            logger_->trace("Calling lua function {}", func);
+            if (!isCreated_) {
+                logger_->warn("Trying to call lua function {}() when the lua environment hasn't been created", func);
+                return {};
+            }
+            sol::protected_function f = env_[func];
+            auto res = f(args...);
+            if (res.valid())
+                return static_cast<T>(res);
+            sol::error err = res;
+            logger_->error("{}() lua call failed with: {}", func, err.what());
+            return {};
+        }
 };
 
 }
